Moving adif formatting functions into field specific formatters
This patch moves all of the formatting functions previously hosted under the
::adif namespace to the ::adif::formatters namespace. Formatter names are also
standardized to be <field-name>.from to indicate they are translating from the
ADIF field to the human readable name.

The patch also introduces automated field formatting in the ::adif::getField
function. If a formatter for the named field exists, it is automatically called
to translate the ADIF value into a human readable value before returning.