Generalizing the formatter code to make it easy to add new enumeration formats.
Many of the ADIF conversions are simple enumerations mapping the value in the
ADIF file to some more human-readable name. This patch simplifies this logic
by adding simple helper functions to create case insensitive lookup tables
and a simple function to perform the lookup.

The logic to create the lookup tables reads the values from separate files,
eliminating multi-line enumerations from needing to be specified in code.

The DXCC and CONT conversion functions are being updated to use these new
helper functions.